+++ src/Application/Products/Commands/DeleteProduct/DeleteProductCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using Application.Common.Interfaces;
 using Application.Common.Messaging;
 using Application.Common.Models;
+using Microsoft.EntityFrameworkCore;
 
 namespace Application.Products.Commands.DeleteProduct
 {
@@ -14,10 +15,18 @@ namespace Application.Products.Commands.DeleteProduct
         }
         public async Task<Result> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
-            var product = await _context.Products.FindAsync(new object[] { request.Id }, cancellationToken);
+            var product = await _context.Products
+                .Include(p => p.Assets)
+                .SingleOrDefaultAsync(p => p.Id == request.Id, cancellationToken);
+
             if (product is null || product.IsDeleted)
             {
                 return Result.Failure(AssetErrors.ProductNotFound);
+            }
+
+            if (product.Assets.Count > 0)
+            {
+                return Result.Failure(AssetErrors.ProductCurrentlyUsed);
             }
 
             _context.Products.Remove(product);
